requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.2
PyMySQL==1.0.2
requests==2.28.2
python-dotenv==1.0.0 